+++ Classes/Controller/Example/FilesController.php
@@ -3,6 +3,7 @@
 namespace Xima\XimaTypo3Recordlist\Controller\Example;
 
 use Xima\XimaTypo3Recordlist\Controller\AbstractBackendController;
+use Xima\XimaTypo3Recordlist\Dto\RecordSource;
 
 class FilesController extends AbstractBackendController
 {
@@ -11,9 +12,9 @@ class FilesController extends AbstractBackendController
         return ['sys_file_metadata'];
     }
 
-    public function getRecordPid(): int
+    protected function getRecordSources(): array
     {
-        return 0;
+        return [new RecordSource(0, includeSubpages: true, depth: 1)];
     }
 
     protected function addNewButtonToModuleTemplate(): void