+++ Classes/Controller/Example/SysCategoryController.php
@@ -3,12 +3,13 @@
 namespace Xima\XimaTypo3Recordlist\Controller\Example;
 
 use Xima\XimaTypo3Recordlist\Controller\AbstractBackendController;
+use Xima\XimaTypo3Recordlist\Dto\RecordSource;
 
 class SysCategoryController extends AbstractBackendController
 {
-    public function getRecordPid(): int
+    protected function getRecordSources(): array
     {
-        return 17;
+        return [new RecordSource(17, includeSubpages: true, depth: 1)];
     }
 
     public function getTableNames(): array